+++ integration_tests/api/replication_api_test.py
@@ -25,6 +25,7 @@ def test__create_replication_ok(base_url, session, bucket_name, replication_name
             "exclude": {"key2": "value2"},
             "each_n": 10,
             "each_s": 0.5,
+            "when": {"$eq": ["&key1", "value1"]},
         },
     )
 
@@ -50,6 +51,7 @@ def test__create_replication_ok(base_url, session, bucket_name, replication_name
             "include": {"key1": "value1"},
             "each_n": 10,
             "each_s": 0.5,
+            "when": {"$eq": ["&key1", "value1"]},
         },
     }
 
@@ -118,6 +120,7 @@ def test__update_replication_ok(base_url, session, bucket_name, replication_name
             "each_n": None,
             "each_s": None,
             "src_bucket": bucket_name,
+            "when": None,
         },
     }