+++ reductstore/src/replication/replication_repository.rs
@@ -9,6 +9,7 @@ use crate::replication::proto::{
 };
 use crate::replication::replication_task::ReplicationTask;
 use crate::replication::{ManageReplications, TransactionNotification};
+use crate::storage::query::condition::Parser;
 use crate::storage::storage::Storage;
 use bytes::Bytes;
 use log::{debug, error};
@@ -45,6 +46,7 @@ impl From<ReplicationSettings> for ProtoReplicationSettings {
                 .collect(),
             each_s: settings.each_s.unwrap_or(0.0),
             each_n: settings.each_n.unwrap_or(0),
+            when: settings.when.map(|value| value.to_string()),
         }
     }
 }
@@ -74,6 +76,20 @@ impl From<ProtoReplicationSettings> for ReplicationSettings {
             },
             each_n: if settings.each_n > 0 {
                 Some(settings.each_n)
+            } else {
+                None
+            },
+            when: if let Some(when) = settings.when {
+                match serde_json::from_str(&when) {
+                    Ok(value) => Some(value),
+                    Err(err) => {
+                        error!(
+                            "Failed to parse 'when' field: {} in replication settings: {}",
+                            err, when
+                        );
+                        None
+                    }
+                }
             } else {
                 None
             },
@@ -284,6 +300,17 @@ impl ReplicationRepository {
                 "Source and destination buckets must be different",
             ));
         }
+
+        // check syntax of when condition
+        if let Some(when) = &settings.when {
+            if let Err(e) = Parser::new().parse(when) {
+                return Err(unprocessable_entity!(
+                    "Invalid replication condition: {}",
+                    e
+                ));
+            }
+        }
+
         let replication =
             ReplicationTask::new(name.to_string(), settings, Arc::clone(&self.storage));
         self.replications.insert(name.to_string(), replication);
@@ -378,6 +405,22 @@ mod tests {
                     "Source bucket 'bucket-2' for replication 'test' does not exist"
                 )),
                 "Should not create replication with non existing source bucket"
+            );
+        }
+
+        #[rstest]
+        fn test_replication_with_invalid_when_condition(
+            mut repo: ReplicationRepository,
+            settings: ReplicationSettings,
+        ) {
+            let mut settings = settings;
+            settings.when = Some(serde_json::json!({"$UNKNOWN_OP": ["&x", "y"]}));
+            assert_eq!(
+                repo.create_replication("test", settings),
+                Err(unprocessable_entity!(
+                    "Invalid replication condition: [UnprocessableEntity] Operator '$UNKNOWN_OP' not supported"
+                )),
+                "Should not create replication with invalid when condition"
             );
         }
 
@@ -595,8 +638,8 @@ mod tests {
         #[rstest]
         fn test_from_proto(settings: ReplicationSettings) {
             let proto_settings = ProtoReplicationSettings::from(settings.clone());
-            let settings = ReplicationSettings::from(proto_settings);
-            assert_eq!(settings, settings);
+            let restored_settings = ReplicationSettings::from(proto_settings);
+            assert_eq!(settings, restored_settings);
         }
 
         #[rstest]
@@ -604,8 +647,8 @@ mod tests {
             let mut settings = settings;
             settings.each_n = Some(10);
             let proto_settings = ProtoReplicationSettings::from(settings.clone());
-            let settings = ReplicationSettings::from(proto_settings);
-            assert_eq!(settings, settings);
+            let restored_settings = ReplicationSettings::from(proto_settings);
+            assert_eq!(settings, restored_settings);
         }
 
         #[rstest]
@@ -613,8 +656,26 @@ mod tests {
             let mut settings = settings;
             settings.each_s = Some(10.0);
             let proto_settings = ProtoReplicationSettings::from(settings.clone());
-            let settings = ReplicationSettings::from(proto_settings);
-            assert_eq!(settings, settings);
+            let restored_settings = ReplicationSettings::from(proto_settings);
+            assert_eq!(settings, restored_settings);
+        }
+
+        #[rstest]
+        fn test_from_when_proto(settings: ReplicationSettings) {
+            let mut settings = settings;
+            settings.when = Some(serde_json::json!({"$and": [true, true]}));
+            let proto_settings = ProtoReplicationSettings::from(settings.clone());
+            let restored_settings = ReplicationSettings::from(proto_settings);
+            assert_eq!(settings, restored_settings);
+        }
+
+        #[rstest]
+        fn test_from_when_proto_invalid(settings: ReplicationSettings) {
+            let mut proto_settings = ProtoReplicationSettings::from(settings.clone());
+            proto_settings.when = Some("invalid".to_string());
+
+            let restored_settings = ReplicationSettings::from(proto_settings);
+            assert!(restored_settings.when.is_none());
         }
     }
 
@@ -630,6 +691,7 @@ mod tests {
             exclude: Labels::default(),
             each_n: None,
             each_s: None,
+            when: None,
         }
     }