+++ reductstore/src/storage/query.rs
@@ -2,9 +2,9 @@
 // Licensed under the Business Source License 1.1
 
 pub mod base;
-mod condition;
+pub(crate) mod condition;
 mod continuous;
-pub mod filters;
+pub(crate) mod filters;
 mod historical;
 mod limited;